Training/requirements.txt
biopython==1.83
numpy==1.26.4
scikit-learn==1.3.2
scipy==1.12.0

